+++ tests/Regressors/RegressionTreeTest.php
@@ -31,7 +31,7 @@ class RegressionTreeTest extends TestCase
     {
         $this->generator = new SwissRoll(4., -7., 0., 1., 0.2);
 
-        $this->estimator = new RegressionTree(30, 2, 0., null, 1e-4);
+        $this->estimator = new RegressionTree(30, 3, 0., null, 1e-4);
 
         $this->metric = new RSquared();
     }